Revert "Remove package manager. Move to standard-library"
This reverts commit f9a93469dab43eeacd8dcf5a4db9e6e0d72c6009.
