+++ library/Backends/Michelson/test/Main.hs
@@ -7,11 +7,9 @@ import qualified Test.Parameterisation as Parameterisation
 import qualified Test.Pipeline as Pipeline
 import qualified Test.Tasty as T
 import qualified Test.VStack as VStack
-import Juvix.Library.StdLib (loadStdLibs)
 
 allCheckedTests :: IO T.TestTree
 allCheckedTests = do
-  loadStdLibs
   goldenTests <- Golden.top
   pure $
     T.testGroup
