+++ library/Backends/Plonk/test/Main.hs
@@ -4,11 +4,9 @@ import Juvix.Library
 import qualified Test.Compiler as Compiler
 import qualified Test.Golden as Golden
 import qualified Test.Tasty as T
-import Juvix.Library.StdLib (loadStdLibs)
 
 main :: IO ()
 main = do
-  loadStdLibs
   compilerTests <- Compiler.top
   goldenTests <- Golden.top
   T.defaultMain $
