+++ library/Backends/llvm/test/Main.hs
@@ -1,14 +1,12 @@
 module Main where
 
 import Juvix.Library
-import Juvix.Library.StdLib (loadStdLibs)
 import qualified Test.Golden as Golden
 import qualified Test.Parameterization as Parameterization
 import qualified Test.Tasty as T
 
 main :: IO ()
 main = do
-  loadStdLibs
   goldenTests <- Golden.top
   T.defaultMain $
     T.testGroup "LLVM tests" [goldenTests, Parameterization.top]
