+++ library/Core/test/Main.hs
@@ -8,7 +8,6 @@ import Juvix.Library (IO)
 import qualified Test.Tasty as T
 import Typechecker (coreCheckerEval)
 import qualified Utility
-import Juvix.Library.StdLib (loadStdLibs)
 
 coreTests :: T.TestTree
 coreTests =
@@ -30,6 +29,4 @@ allCheckedTests =
     ]
 
 main :: IO ()
-main = do
-  loadStdLibs
-  T.defaultMain allCheckedTests
+main = T.defaultMain allCheckedTests
