+++ library/Frontend/test/Main.hs
@@ -3,7 +3,7 @@ module Main where
 import Juvix.Library (IO)
 import Parser (allParserTests)
 import qualified Test.Tasty as T
-import Juvix.Library.StdLib (loadStdLibs)
+
 allCheckedTests :: T.TestTree
 allCheckedTests =
   T.testGroup
@@ -11,6 +11,4 @@ allCheckedTests =
     [allParserTests]
 
 main :: IO ()
-main = do
-  loadStdLibs
-  T.defaultMain allCheckedTests
+main = T.defaultMain allCheckedTests
