+++ library/Pipeline/test/Main.hs
@@ -4,7 +4,6 @@ import Juvix.Library (IO)
 import qualified Test.RecGroups as RecGroups
 import qualified Test.Tasty as T
 import qualified Test.ToSexp as ToSexp
-import Juvix.Library.StdLib (loadStdLibs)
 
 allCheckedTests :: T.TestTree
 allCheckedTests =
@@ -13,6 +12,4 @@ allCheckedTests =
     [RecGroups.top, ToSexp.top]
 
 main :: IO ()
-main = do
-  loadStdLibs
-  T.defaultMain allCheckedTests
+main = T.defaultMain allCheckedTests
