+++ library/Playground/HTTP/test/Main.hs
@@ -1,9 +1,6 @@
 module Main where
 
 import Juvix.Library
-import Juvix.Library.StdLib (loadStdLibs)
 
 main :: IO ()
-main = do
-    loadStdLibs
-    pure ()
+main = pure ()
