+++ library/Sexp/test/Main.hs
@@ -5,7 +5,7 @@ import qualified Sexp
 import qualified Sexp.Parser
 import qualified Sexp.SimplifiedPasses
 import qualified Test.Tasty as T
-import Juvix.Library.StdLib (loadStdLibs)
+
 allCheckedTests :: T.TestTree
 allCheckedTests =
   T.testGroup
@@ -16,6 +16,4 @@ allCheckedTests =
     ]
 
 main :: IO ()
-main = do
-  loadStdLibs
-  T.defaultMain allCheckedTests
+main = T.defaultMain allCheckedTests
