+++ library/Test/PackageManager/package.yaml
@@ -1,0 +1,79 @@
+name:                package-manager
+version:             0.1.0.0
+github:              "heliaxdev/Juvix"
+license:             GPL-3
+author:              "Heliax AG <hello@heliax.dev>"
+maintainer:          "Heliax AG <hello@heliax.dev>"
+copyright:           "2020 Heliax AG"
+
+dependencies:
+- aeson
+- base >= 4.7 && < 5
+- bytestring
+- directory
+- filepath
+- generic-lens
+- lens
+- lens-aeson
+- pretty-simple
+- standard-library
+- text
+- unordered-containers
+- wreq
+
+
+default-extensions:
+  - NoImplicitPrelude
+  - OverloadedStrings
+  - NoMonomorphismRestriction
+  - RankNTypes
+  - LambdaCase
+  - UnicodeSyntax
+  - GADTs
+  - DerivingVia
+  - MultiParamTypeClasses
+  - MultiWayIf
+  - DataKinds
+  - TypeApplications
+  - ConstraintKinds
+  - PatternSynonyms
+  - FlexibleContexts
+  - FlexibleInstances
+  - QuasiQuotes
+  - TemplateHaskell
+  - TypeFamilies
+  - NamedFieldPuns
+  - DisambiguateRecordFields
+  - TupleSections
+  - DeriveGeneric
+  - DeriveDataTypeable
+  - GeneralizedNewtypeDeriving
+  - StandaloneDeriving
+  - BlockArguments
+  - FunctionalDependencies
+
+
+ghc-options:
+  - -ferror-spans
+  - -Wall
+  - -fno-warn-orphans
+  - -fno-warn-name-shadowing
+  - -fno-warn-missing-pattern-synonym-signatures
+  - -j
+  - -static
+  - -fwrite-ide-info
+
+
+library:
+  source-dirs: src
+
+tests:
+  data-structures-test:
+    main:                Spec.hs
+    source-dirs:         test
+    ghc-options:
+    - -threaded
+    - -rtsopts
+    - -with-rtsopts=-N
+    dependencies:
+    - data-structures
