+++ library/Test/PackageManager/src/Juvix/PackageManager/StdLib.hs
@@ -1,13 +1,13 @@
 {-# LANGUAGE ViewPatterns #-}
-module Juvix.Library.StdLib where
+module Juvix.PackageManager.StdLib where
 
-import Protolude
+import Juvix.Library
 import System.Directory
 import Text.Pretty.Simple
 import qualified Data.ByteString.Lazy as BL
 import Control.Lens
 import qualified Network.Wreq as Wreq
-import Data.Aeson.Lens (key, _Object)
+import Data.Aeson.Lens (key)
 import qualified Data.HashMap.Strict as Map
 import System.FilePath.Posix
 
@@ -51,6 +51,6 @@ localStdLibs = do
 loadStdLibs :: IO ()
 loadStdLibs = do
   -- TODO: How long do we want to cache this?
+  success <- localStdLibs
   -- TODO: Checksum
-  -- success <- localStdLibs
   unless success downloadStdLibs
