+++ library/Test/PackageManager/stack.yaml
@@ -1,0 +1,26 @@
+resolver: lts-17.3
+
+packages:
+- .
+- ../../StandardLibrary
+
+extra-deps:
+
+########################
+# General Dependencies #
+########################
+- capability-0.4.0.0@sha256:d86d85a1691ef0165c77c47ea72eac75c99d21fb82947efe8b2f758991cf1837,3345
+- github: jyp/prettiest
+  commit: e5ce6cd6b4da71860c3d97da84bed4a827fa00ef
+- git: https://github.com/serokell/galois-field.git
+  commit: 576ba98ec947370835a1f308895037c7aa7f8b71
+- git: https://github.com/serokell/elliptic-curve.git
+  commit: b8a3d0cf8f7bacfed77dc3b697f5d08bd33396a8
+
+#####################################
+# Standard Library Extra Dependency #
+#####################################
+- github: phile314/tasty-silver
+  commit: f1f90ac3113cd445e2a7ade43ebb29f0db38ab9b
+- tasty-1.4.1@sha256:69e90e965543faf0fc2c8e486d6c1d8cf81fd108e2c4541234c41490f392f94f,2638
+
