+++ package.yaml
@@ -115,3 +115,4 @@ executables:
       - juvix
       - pretty-simple
       - optparse-applicative
+      - package-manager
