+++ scripts/yaml-generator/src/config.lisp
@@ -328,6 +328,14 @@ common ones to include"
    :extra-deps (list (general-dependencies)
                      *standard-library-extra-deps*)))
 
+(defparameter *package-manager*
+  (make-stack-yaml
+   :name     "Test/PackageManager"
+   :path-to-other "../../"
+   :packages   (list *standard-library*)
+   :extra-deps (list (general-dependencies)
+                     *standard-library-extra-deps*)))
+
 (defparameter *core*
   (make-stack-yaml
    :name       "Core"
@@ -493,7 +501,8 @@ common ones to include"
                    *witch*
                    *context*
                    *sexp*
-                   *data-structures*)
+                   *data-structures*
+                   *package-manager*)
    :path-to-other "./library/"
    :extra-deps
    (cons *servant-deps* (cons *llvm-hs-deps* (big-dep-list)))
