+++ scripts/yaml-generator/src/main.lisp
@@ -24,4 +24,5 @@
   (generate-yaml-file *easy*             "library/Playground/Easy/stack.yaml")
   (generate-yaml-file *http*             "library/Playground/HTTP/stack.yaml")
   (generate-yaml-file *data-structures*   "library/Test/DataStructures/stack.yaml")
+  (generate-yaml-file *package-manager*   "library/Test/PackageManager/stack.yaml")
   (generate-yaml-file *juvix*            "stack.yaml"))
